requirements.txt
pyserial
loguru
